requirements.txt
streamlit
requests
python-dotenv
sentence-transformers
numpy
pillow
scikit-learn
google-generativeai